chore: multiple fixes - game card not showing background - links highlighted - loading in review detail - favorite in selecting genres

+++ frontend/src/app/reviews/feature/review-detail/review-detail.component.ts
@@ -1,36 +1,19 @@
-import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
-import { ReviewsService } from '../../../shared/data-access/reviews/reviews.service';
-import {
-  Feedback,
-  Review,
-} from '../../../shared/data-access/reviews/review.class';
-import {
-  BehaviorSubject,
-  combineLatest,
-  map,
-  Observable,
-  of, startWith,
-  switchMap,
-  tap,
-} from 'rxjs';
-import { AuthenticationService } from '../../../shared/data-access/authentication/authentication.service';
-import {
-  DifficultyToLocale,
-  Role,
-} from '../../../shared/data-access/shared.enums';
-import { environment } from '../../../../environments/environment';
-import { MatDialog, MatDialogModule } from '@angular/material/dialog';
-import { SharedModule } from '../../../shared/shared.module';
-import { FormsModule } from '@angular/forms';
-import { MatRadioModule } from '@angular/material/radio';
-import { NgForOf } from '@angular/common';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { TranslateService } from '@ngx-translate/core';
-import {
-  getMessageFromReason,
-  ReportReason,
-} from '../../../shared/data-access/reports/reports.class';
+import {ChangeDetectionStrategy, Component, OnInit} from '@angular/core';
+import {ActivatedRoute, Router} from '@angular/router';
+import {ReviewsService} from '../../../shared/data-access/reviews/reviews.service';
+import {Feedback, Review,} from '../../../shared/data-access/reviews/review.class';
+import {BehaviorSubject, combineLatest, distinctUntilChanged, map, Observable, switchMap, tap,} from 'rxjs';
+import {AuthenticationService} from '../../../shared/data-access/authentication/authentication.service';
+import {DifficultyToLocale, Role,} from '../../../shared/data-access/shared.enums';
+import {environment} from '../../../../environments/environment';
+import {MatDialog, MatDialogModule} from '@angular/material/dialog';
+import {SharedModule} from '../../../shared/shared.module';
+import {FormsModule} from '@angular/forms';
+import {MatRadioModule} from '@angular/material/radio';
+import {NgForOf} from '@angular/common';
+import {MatSnackBar} from '@angular/material/snack-bar';
+import {TranslateService} from '@ngx-translate/core';
+import {getMessageFromReason, ReportReason,} from '../../../shared/data-access/reports/reports.class';
 
 @Component({
   selector: 'app-review-detail',
@@ -49,54 +32,47 @@ export class ReviewDetailComponent implements OnInit {
     private readonly translate: TranslateService
   ) {}
 
-  loading$ = new BehaviorSubject<boolean>(false);
+  loading = true;
 
   review$: Observable<Review> = this.route.paramMap
+    .pipe(map((params) => params.get('id')))
+    .pipe(distinctUntilChanged())
     .pipe(
-      tap(() => this.loading$.next(true)),
-      switchMap(params => {
+      tap(() => this.loading = true),
+      switchMap(id => {
         return this.reviewsService.getReviewById(
-          `${environment.API_ENDPOINT}/reviews/${params.get('id')}`
+          `${environment.API_ENDPOINT}/reviews/${id}`
         );
       }),
-      tap(() => this.loading$.next(false))
+      tap(() => this.loading = false)
     );
 
-  loadingReview$: Observable<{ review: Review | null, loading: boolean }> = combineLatest({ review: this.review$.pipe(startWith(null)), loading: this.loading$ });
+  loadedReview$ = new BehaviorSubject<Review | null>(null);
 
   currentUser$ = this.authService.getLoggedUser();
-  userAndReview$ = combineLatest([this.currentUser$, this.review$]);
-  reviewFeedback$: Observable<Feedback> = of();
+  userAndReview$ = combineLatest([this.currentUser$, this.loadedReview$]);
+
+  reviewFeedback$ = new BehaviorSubject<Feedback | null>(null);
   canEdit$ = this.userAndReview$.pipe(
     map(([user, review]) => {
-      return user != null && user.id === review.authorId;
+      return user != null && user.id === review?.authorId;
     })
   );
   canReport$ = this.userAndReview$.pipe(
     map(([user, review]) => {
-      return user != null && user.id !== review.authorId;
+      return user != null && user.id !== review?.authorId;
     })
   );
   canDelete$ = this.userAndReview$.pipe(
     map(([user, review]) => {
       return (
         user != null &&
-        (user.id === review.authorId || user.role === Role.MODERATOR)
+        (user.id === review?.authorId || user.role === Role.MODERATOR)
       );
     })
   );
 
   ngOnInit() {
-    this.userAndReview$.subscribe(([user, review]) => {
-      if (!user || !review.links.feedback) {
-        this.reviewFeedback$ = of(new Feedback(null));
-        return;
-      }
-      this.reviewFeedback$ = this.reviewsService.getReviewFeedbackFromUser(
-        review.links.feedback
-      );
-    });
-
     this.route.paramMap.subscribe(params => {
       const id = params.get('id');
       if (!(id && Number.isInteger(parseInt(id)) && parseInt(id) > 0)) {
@@ -107,6 +83,18 @@ export class ReviewDetailComponent implements OnInit {
       error: () => {
         this.router.navigate(['/errors/not-found']);
       },
+      next: (review) => {
+        this.loadedReview$.next(review);
+        this.currentUser$.subscribe((user) => {
+          if (user !== null && review.links.feedback) {
+            this.reviewsService.getReviewFeedbackFromUser(
+              review.links.feedback
+            ).subscribe((feedback) => this.reviewFeedback$.next(feedback));
+          } else {
+            this.reviewFeedback$.next(null);
+          }
+        })
+      }
     });
   }
 
@@ -117,6 +105,7 @@ export class ReviewDetailComponent implements OnInit {
         //confirmed
         console.log('deleting');
         this.review$.subscribe(review => {
+          if (!review) return;
           this.reviewsService.deleteReview(review.links.self).subscribe({
             error: () => {
               this.snackBar.open(
@@ -150,7 +139,7 @@ export class ReviewDetailComponent implements OnInit {
     dialogRef.afterClosed().subscribe(result => {
       if (result != null) {
         this.review$.subscribe(review => {
-          if (!review.links.report) return;
+          if (!review?.links.report) return;
           this.reviewsService
             .reportReview(review.links.report, review.id, result)
             .subscribe({
